refactor(userspace-dp): split event_stream/tests.rs into per-concern files (#4664)
Pure test-code motion. The event_stream tests lived in a single
2313-line tests.rs loaded via `#[path = "tests.rs"] mod tests;`, over the
modularity-discipline LOC threshold. Split it into a tests/ subdir with
per-concern sibling submodules, matching the #4651/#4665 convention (a
tests/mod.rs holding the shared header imports + cross-concern fixtures,
each concern reached via `use super::*`).

Concern files and their test counts:
  rt_flow.rs         (13) RT_FLOW SESSION_CREATE/CLOSE emit + monotonic
                          -> wall-clock conversions
  replay_budget.rs   (13) seq allocation, replay-buffer trim/eviction,
                          telemetry queue budget, lossless seq accounting
  control_frames.rs  (10) ACK watermark validation (#2959), partial/
                          multi-frame reads, bounded control payloads (#2879)
  drain.rs            (9) DrainRequest fence filtering (#2876/#2882),
                          stuck-reader backpressure (#2877/#2883),
                          paused-eviction drain poisoning (#2875)
  backpressure.rs     (7) channel + write-backlog shedding/stall counting,
                          paused drain, lossless capacity waiting

The shared `use` header and the three cross-concern fixtures
(build_raw_ack_frame, test_dataplane_event, test_close_delta) live in
tests/mod.rs; concern-local helpers travel with their tests. The parent
event_stream/mod.rs declaration changed from
`#[path = "tests.rs"] mod tests;` to plain `mod tests;` (resolving to
tests/mod.rs) — the only production-file line touched, and it sits inside
the existing `#[cfg(test)]` gate. The stale file-header comment that
documented the old `#[path]` layout was dropped.

No production code and no test logic changed. Every test/helper body
moved byte-identical (extracted via an exact line-range partition of the
original file). A code-line multiset check confirms zero original code
lines are missing and the only new lines are the five `mod X;` decls plus
the five per-file `use super::*;` wiring lines.

Validation: cargo build OK; `cargo test --release -- --test-threads=1`
FULL suite 3854 passed / 0 failed / 2 ignored, identical to the pre-split
pristine baseline; `cargo test --release event_stream` green;
event_stream::tests count 52 -> 52 (13+13+10+9+7). The lone pre-existing
`unused variable: handle` warning is carried over verbatim; no new
warnings introduced.
